secure revision: Add warning about the workflow for secure revisions.
Secure revisions are managed in Bugzilla. Users should be notified about
this behaviour.

To place a warning info on the revision's page a new plugin is created.
Phabricator's code is patched so it will call the plugin and place it
above the "Details" panel and any potential warnings from Phabricator.

Plugin is extendable - we can add warnings, to the page by modifying the
plugin only.