suite: Adding Mozilla MOTD CSS File
MOTD info window is displayed under the resize bar if file tree is
switched on. This patch adds a CSS file which is fixing rhis issue.

**WARNING:**
There might be a need to do an administrative task:

If an Exception of missing celerity resource is thrown - run
$ docker-compose exec phabricator sh
$ ./phabricator/bin/celery map

The exception is nondeterministic. I've run it once in the suite on an image
created in `phabricator-extensions` dir, removed images and the Exception
was not there anymore.
Please keep an eye on this to avoid dealing with this problem on
dev/production.